Remove purposeless return values from Adapter and Filesystem interfaces
So far write, delete and rename commands return values: the number
of written bytes for write, and a boolean value for the two others.

The Local adapter was the first adapter to be introduced (see eed10e367),
and I think the Adapter interface has been designed like that to try
to mimic low-level file functions but it looks looks a bad design
decision right now because it forces most adapters to recompute the size
of the written content locally. If we want to check that the size of the
written content matches the size of the content passed to the filesystem
we have to recompute it locally too. That's crazy.

Return values for low-level abstractions are purposeful, but I don't
see any good reason for higher-level abstraction. Plus it's better in OO
to not return any value when something happens. So propose to remove
return values from write, delete and rename OO commands.

Right now the work has been done for AwsS3 only.